+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 ## [Unreleased]
 
+## [0.2.0]
+### Changed
+- Changed the file chooser dialog in the 'Open Index' and 'Create Index' dialogs to an IntelliJ one to provide
+a smoother user experience.
+
 ## [0.1.0]
 ### Added
 - Integrated Lucene's Luke application into this IDE plugin. Code changes from Luke are applied up to version 9.10.0.